RDCC-120 Refactored integration test code
